requirements.txt
# run this: 
# pip install -r requirements.txt
# or
# pip install --upgrade --upgrade-strategy eager -r requirements.txt

flask
aiohttp
aiosignal
alembic
annotated-types
anyio
appdirs
asgiref
async-timeout
attrs
backoff
bcrypt
beautifulsoup4
boto3
botocore
Brotli
build
cachetools
certifi
cffi
charset-normalizer
chroma-hnswlib
chromadb
click
cohere
colorama
coloredlogs
crewai
crewai-tools
cryptography
dataclasses-json
decorator
Deprecated
deprecation
distro
docstring-parser
docx2txt
embedchain
exceptiongroup
fastapi
fastavro
filelock
flatbuffers
frozenlist
fsspec
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-core
google-cloud-resource-manager
google-cloud-storage
google-crc32c
google-generativeai
google-resumable-media
googleapis-common-protos
gptcache
greenlet
grpc-google-iam-v1
grpcio
grpcio-status
h11
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib-metadata
importlib_resources
iniconfig
instructor
jmespath
jsonpatch
jsonpointer
kubernetes
lancedb
langchain
langchain-cohere
langchain-community
langchain-core
langchain-google-genai
langchain-openai
langchain-text-splitters
langsmith
load-dotenv
Mako
markdown-it-py
MarkupSafe
marshmallow
mdurl
mmh3
monotonic
mpmath
multidict
mutagen
mypy-extensions
nodeenv
numpy
oauthlib
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
outcome
overrides
packaging
pluggy
posthog
proto-plus
protobuf
pulsar-client
py
pyarrow
pyasn1
pyasn1_modules
pycparser
pycryptodomex
pydantic
pydantic_core
PyGithub
Pygments
PyJWT
pylance
PyNaCl
pyparsing
pypdf
PyPika
pyproject_hooks
pyreadline3
pyright
pysbd
PySocks
pytest
python-dateutil
python-dotenv
pytube
PyYAML
ratelimiter
regex
requests
requests-oauthlib
retry
rich
rsa
s3transfer
schema
selenium
semver
shapely
six
smmap
sniffio
sortedcontainers
soupsieve
SQLAlchemy
starlette
sympy
tenacity
tiktoken
tokenizers
tomli
tqdm
trio
trio-websocket
typer
types-requests
typing-inspect
typing_extensions
uritemplate
urllib3
uvicorn
watchfiles
websocket-client
websockets
wrapt
wsproto
yarl
youtube-transcript-api
yt-dlp
zipp
